# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API de Gestão Escolar",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/appointments/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Filtro de busca",
            "in": "query",
            "required": false,
            "type": "object",
            "schema": {
              "$ref": "#/components/schemas/appointment"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/appointment"
              }
            }
          }
        }
      }
    },
    "/appointments/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "specialty": {
                    "example": "any"
                  },
                  "comments": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "student": {
                    "example": "any"
                  },
                  "professional": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Filtro de busca",
            "in": "query",
            "required": false,
            "type": "object",
            "schema": {
              "$ref": "#/components/schemas/event"
            },
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/event"
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "example": "any"
                  },
                  "comments": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "appointment": {
        "type": "object",
        "properties": {
          "specialty": {
            "type": "string",
            "example": "Fisioterapeuta"
          },
          "comments": {
            "type": "string",
            "example": "Realizar sessão"
          },
          "date": {
            "type": "string",
            "example": "2023-08-15 16:00:00"
          },
          "student": {
            "type": "string",
            "example": "Bingo Heeler"
          },
          "professional": {
            "type": "string",
            "example": "Winton Blake"
          }
        },
        "required": [
          "date",
          "student",
          "professional"
        ]
      },
...
      "event": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Palestra bem viver com saúde"
          },
          "comments": {
            "type": "string",
            "example": "Profissionais de saúde da Unesc"
          },
          "date": {
            "type": "string",
            "example": "2023-08-15 16:00:00"
          }
        },
        "required": [
          "description",
          "date"
        ]
      }
    }
  }
}